refactor: relocate misfiled message helpers out of ui/ (#218)
Closes residual core -> @/ui/messages coupling by moving message-builder
modules that were never UI-layer concerns:

- src/ui/messages/debug.ts -> src/daemon/messages.ts
  Thin wrappers returning daemon diagnostic strings. Used by worker,
  startSession, lifecycle/*. Belongs with daemon code, not UI.

- src/ui/messages/errors.ts -> src/errors/messages.ts
  Error prose builders (scriptExecutionError, keyPressFailedError,
  noFormsFoundError, etc.). Belongs next to CommandError, not UI.

Also inline the two trivial happy-path helpers in launcher.ts
(chromeLaunchSuccessMessage, chromeUserDataDirMessage — each a one-line
string builder) and drop them from ui/messages/chrome.ts.

Net: remaining @/ui/messages imports in src/{connection,telemetry,daemon,
session,cdp,runtime,ipc,errors} are only formatChromeIssue/
formatChromeNotice in daemon/worker.ts — which IS the UI boundary where
structured issues become user-visible prose. That's the intended
coupling, not the residual debt.
