+++ src/commands/dom/DomElementResolver.ts
@@ -27,9 +27,9 @@
  */
 
 import { CommandError } from '@/errors/index.js';
+import { elementAtIndexNotFoundError, indexOutOfRangeError } from '@/errors/messages.js';
 import { QueryCacheManager } from '@/session/QueryCacheManager.js';
 import { createLogger } from '@/ui/logging/index.js';
-import { elementAtIndexNotFoundError, indexOutOfRangeError } from '@/ui/messages/errors.js';
 import { EXIT_CODES } from '@/utils/exitCodes.js';
 
 const log = createLogger('dom');
