+++ src/commands/dom/get.ts
@@ -21,9 +21,9 @@ import {
 import { runCommand } from '@/commands/shared/CommandRunner.js';
 import type { DomGetCommandOptions } from '@/commands/shared/optionTypes.js';
 import { CommandError } from '@/errors/index.js';
+import { elementAtIndexNotFoundError, noNodesFoundError } from '@/errors/messages.js';
 import { resolveA11yNode } from '@/telemetry/a11y.js';
 import { formatDomGet } from '@/ui/formatters/dom.js';
-import { elementAtIndexNotFoundError, noNodesFoundError } from '@/ui/messages/errors.js';
 import { EXIT_CODES } from '@/utils/exitCodes.js';
 import { filterDefined } from '@/utils/objects.js';
