+++ src/commands/dom/helpers/screenshot.ts
@@ -14,14 +14,14 @@ import {
 import { CDPConnectionError } from '@/connection/errors.js';
 import type { Protocol } from '@/connection/typed-cdp.js';
 import { CommandError } from '@/errors/index.js';
-import { callCDP } from '@/ipc/client.js';
-import type { ScreenshotResult, ScreenshotOptions, ElementBounds } from '@/types.js';
-import { createLogger } from '@/ui/logging/index.js';
 import {
   noNodesFoundError,
   elementNotVisibleError,
   elementZeroDimensionsError,
-} from '@/ui/messages/errors.js';
+} from '@/errors/messages.js';
+import { callCDP } from '@/ipc/client.js';
+import type { ScreenshotResult, ScreenshotOptions, ElementBounds } from '@/types.js';
+import { createLogger } from '@/ui/logging/index.js';
 import { EXIT_CODES } from '@/utils/exitCodes.js';
 
 const log = createLogger('dom');
