+++ src/commands/network/har.ts
@@ -14,10 +14,10 @@ import { runCommand } from '@/commands/shared/CommandRunner.js';
 import { jsonOption } from '@/commands/shared/commonOptions.js';
 import type { NetworkHarCommandOptions } from '@/commands/shared/optionTypes.js';
 import { CommandError } from '@/errors/index.js';
+import { operationFailedError } from '@/errors/messages.js';
 import { getSessionFilePath } from '@/session/paths.js';
 import { applyFilters, parseFilterString, validateFilterString } from '@/telemetry/filterDsl.js';
 import { buildHAR } from '@/telemetry/har/builder.js';
-import { operationFailedError } from '@/ui/messages/errors.js';
 import { AtomicFileWriter } from '@/utils/atomicFile.js';
 import { EXIT_CODES } from '@/utils/exitCodes.js';
 import { VERSION } from '@/utils/version.js';
