+++ src/daemon/worker/collectors.ts
@@ -3,9 +3,9 @@ import type { TelemetryPlugin } from './plugins.js';
 import type { WorkerConfig } from './types.js';
 
 import type { CDPConnection } from '@/connection/cdp.js';
+import { workerActivatingCollector, workerCollectorsActivated } from '@/daemon/messages.js';
 import type { CleanupFunction, TelemetryType } from '@/types';
 import type { Logger } from '@/ui/logging/index.js';
-import { workerActivatingCollector, workerCollectorsActivated } from '@/ui/messages/debug.js';
 
 import { getRegisteredTelemetryPlugins, shouldActivatePlugin } from './plugins.js';
