+++ src/runtime/dom/evalHelpers.ts
@@ -1,7 +1,7 @@
 import type { CDPConnection } from '@/connection/cdp.js';
 import type { Protocol } from '@/connection/typed-cdp.js';
 import { CommandError } from '@/errors/index.js';
-import { scriptExecutionError } from '@/ui/messages/errors.js';
+import { scriptExecutionError } from '@/errors/messages.js';
 import { EXIT_CODES } from '@/utils/exitCodes.js';
 
 /**
