+++ src/runtime/dom/formFillHelpers/pressKey.ts
@@ -7,9 +7,9 @@
 import type { CDPConnection } from '@/connection/cdp.js';
 import type { Protocol } from '@/connection/typed-cdp.js';
 import { CommandError } from '@/errors/index.js';
+import { keyPressFailedError, operationFailedError } from '@/errors/messages.js';
 import { escapeSelectorForJS } from '@/runtime/dom/formFillHelpers/shared.js';
 import { getKeyDefinition, parseModifiers, type KeyDefinition } from '@/runtime/dom/keyMapping.js';
-import { keyPressFailedError, operationFailedError } from '@/ui/messages/errors.js';
 import { EXIT_CODES } from '@/utils/exitCodes.js';
 
 /**
